chore(tests): add failing testcase for reapplying GlobalJobDslSecurityConfiguration after restart
- fix typo
