+++ src/test/resources/org/jenkinsci/plugins/casc/JobDslGlobalSecurityConfigurationTest.yml
